make benchmark less stupid
Previously we were hashing over all the buckets, meaning the benchmark
times depended on how many nodes were being processed.  This normalized
the benchmark to depending only on b.N (as it should).  It's now clear
that the lookup function is now roughly constant w/r/t to number of
keys, and only depends on memory latency at the higher node numbers.

http://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go